separate arg validation int its own source file
